datalake: rename arrow_schema_translator to arrow_translator
The translator handles both schemas and data items. Renaming it to
better match how it's used.
