+++ src/v/datalake/CMakeLists.txt
@@ -6,7 +6,7 @@ find_package(Protobuf REQUIRED)
 v_cc_library(
   NAME datalake
   SRCS
-    arrow_schema_translator.cc
+    arrow_translator.cc
     protobuf_to_arrow_converter.cc
     proto_to_arrow_scalar.cc
     proto_to_arrow_struct.cc
